+++ HISTORY.md
@@ -2,7 +2,10 @@
 
 # Version 1.8.0
   **Changes to Bot:**
-  - Remove **'No Result found'** message after **5 seconds**.
+  - Remove **'No Result found'** message after **4 seconds**.
+  - Changed 'Search Result' and 'No Result Found' messages.
+  - Moved MessageEmbed 'Descriptions' to 'Title'.
+  - Set 'Chapter per Page' results from **5** to **7**.
   
   **Changes to [KanoKariOCR.json] File:**
   - Added **Ch. 103 - 111** -> Contents.